+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ### Fixed
 
+- #5681 Reset la pagination quand on change de mode ou de sens de tri
 - #5528 Éviter des collisions avec les noms de cache des fichiers distants
 - #5669 Contenu principal toujours en pleine largeur sur petits écrans